refactor: extra the handle logic success to a hook for reusability
