+++ src/app/(card)/admin/autentificare/page.tsx
@@ -2,35 +2,21 @@
 import Button from '@/components/Button'
 import Input from '@/components/Input'
 import { useRouter, useSearchParams } from 'next/navigation'
-import ms from 'ms'
-import { setCookie } from 'cookies-next'
 import useSubmitForm from '@/utils/useSubmitForm'
 import { useEffect } from 'react'
+import { useHandleAuthSuccess } from '@/utils/useHandleAuthSuccess'
 
 // Accepted query parameters:
 // error?: string (an error to show to the user)
 // next?: string (the url to redirect the users to after logining in)
 
-export default function RegisterPage() {
+export default function LoginPage() {
     const router = useRouter()
     const params = useSearchParams()
-
-    const handleSuccess = (data: { name: string; email: string }) => {
-        const date30DaysInTheFuture = new Date(Date.now() + ms('30 days'))
-
-        setCookie('user.name', data.name, {
-            expires: date30DaysInTheFuture,
-        })
-        setCookie('user.email', data.email, {
-            expires: date30DaysInTheFuture,
-        })
-
-        const nextRoute = params.get('next')
-        router.push(typeof nextRoute === 'string' ? nextRoute : '/admin')
-    }
+    const handleAuthSuccess = useHandleAuthSuccess()
 
     const { error, setError, fieldErrors, handleOnSubmit, loading } =
-        useSubmitForm('/users/login', handleSuccess)
+        useSubmitForm('/users/login', handleAuthSuccess)
 
     useEffect(() => {
         const err = params.get('error')
